WinClient-16 Move Interfaces and DTOs to common project
