+++ SysHv/SysHv.Client.Common/DTOs/HardwareInfoDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace SysHv.Client.Common.DTOs
+{
+    public class HardwareInfoDTO
+    {
+
+    }
+}
