+++ SysHv/SysHv.Client.Common/DTOs/MotherBoardDTO.cs
@@ -1,0 +1,9 @@
+﻿namespace SysHv.Client.Common.DTOs
+{
+    public class MotherBoardDTO
+    {
+        public string Product { get; set; }
+
+        public string Manufacturer { get; set; }
+    }
+}
