+++ SysHv/SysHv.Client.Common/DTOs/ProcessorDTO.cs
@@ -1,0 +1,11 @@
+﻿namespace SysHv.Client.Common.DTOs
+{
+    public class ProcessorDTO
+    {
+        public string Id { get; set; }
+
+        public string Manufacturer { get; set; }
+
+        public string CurrentClockSpeed { get; set; }
+    }
+}
