+++ SysHv/SysHv.Client.Common/DTOs/RamDTO.cs
@@ -1,0 +1,11 @@
+﻿namespace SysHv.Client.Common.DTOs
+{
+    public class RamDTO
+    {
+        public string Id { get; set; }
+
+        public string Capacity { get; set; } = "";
+
+        public string MemoryType { get; set; } = "";
+    }
+}
