+++ SysHv/SysHv.Client.Common/Interfaces/IGatherer.cs
@@ -1,4 +1,4 @@
-﻿namespace SysHv.Client.TopShelfService.Interfaces
+﻿namespace SysHv.Client.Common.Interfaces
 {
     public interface IGatherer
     {
