--- SysHv/SysHv.Client.TopShelfService/DTOs/HardwareInfoDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SysHv.Client.TopShelfService.DTOs
-{
-    public class HardwareInfoDTO
-    {
-
-    }
-}
