--- SysHv/SysHv.Client.TopShelfService/DTOs/MotherBoardDTO.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SysHv.Client.TopShelfService.DTOs
-{
-    class MotherBoardDTO
-    {
-        public string Product { get; set; }
-
-        public string Manufacturer { get; set; }
-    }
-}
