--- SysHv/SysHv.Client.TopShelfService/DTOs/ProcessorDTO.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SysHv.Client.TopShelfService.DTOs
-{
-    class ProcessorDTO
-    {
-        public string Id { get; set; }
-
-        public string Manufacturer { get; set; }
-
-        public string CurrentClockSpeed { get; set; }
-    }
-}
