+++ SysHv/SysHv.sln
@@ -9,6 +9,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "client", "client", "{C27054
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.TopShelfService", "SysHv.Client.TopShelfService\SysHv.Client.TopShelfService.csproj", "{982E7E19-5E22-4948-A311-2D5FCAF02EFB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SysHv.Client.Common", "SysHv.Client.Common\SysHv.Client.Common.csproj", "{F0699901-1F2A-4395-B568-B5725FB1C135}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,14 @@ Global
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Release|x64.ActiveCfg = Release|Any CPU
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB}.Release|x64.Build.0 = Release|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Debug|x64.Build.0 = Debug|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Release|x64.ActiveCfg = Release|Any CPU
+		{F0699901-1F2A-4395-B568-B5725FB1C135}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -40,6 +50,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{EFF890AB-FECB-498C-8555-B701C59A4AAE} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
 		{982E7E19-5E22-4948-A311-2D5FCAF02EFB} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
+		{F0699901-1F2A-4395-B568-B5725FB1C135} = {C2705447-0039-4E82-A99C-9D4B6F854C6B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D07E76CC-F6A4-46F4-9E0F-F0EF0952E947}
